+++ scripts/sensor/_02_decoding_full_epochs.py
@@ -1,0 +1,198 @@
+"""
+====================================================
+Decoding pairs of conditions based on entire epochs.
+====================================================
+
+Entire epochs are fed into a logistic regression model. Decoding performance
+then tells how well the classifier could predict which epoch belongs to
+which condition.
+"""
+
+###############################################################################
+# Let us first import the libraries
+
+import os.path as op
+import logging
+from typing import Optional
+import itertools
+from types import SimpleNamespace
+
+import numpy as np
+import pandas as pd
+from scipy.io import savemat
+
+from sklearn.model_selection import cross_val_score
+from sklearn.pipeline import make_pipeline
+from sklearn.model_selection import StratifiedKFold
+
+import mne
+from mne.decoding import Scaler, Vectorizer
+from mne_bids import BIDSPath
+
+import config
+from config import gen_log_kwargs, on_error, failsafe_run, LogReg
+
+
+logger = logging.getLogger('mne-bids-pipeline')
+
+
+@failsafe_run(on_error=on_error, script_path=__file__)
+def run_epochs_decoding(*, cfg, subject, condition1, condition2, session=None):
+    msg = f'Contrasting conditions: {condition1} – {condition2}'
+    logger.info(**gen_log_kwargs(message=msg, subject=subject,
+                                 session=session))
+
+    fname_epochs = BIDSPath(subject=subject,
+                            session=session,
+                            task=cfg.task,
+                            acquisition=cfg.acq,
+                            run=None,
+                            recording=cfg.rec,
+                            space=cfg.space,
+                            suffix='epo',
+                            extension='.fif',
+                            datatype=cfg.datatype,
+                            root=cfg.deriv_root,
+                            check=False)
+
+    epochs = mne.read_epochs(fname_epochs)
+    if cfg.analyze_channels:
+        # We special-case the average reference here to work around a situation
+        # where e.g. `analyze_channels` might contain only a single channel:
+        # `concatenate_epochs` below will then fail when trying to create /
+        # apply the projection. We can avoid this by removing an existing
+        # average reference projection here, and applying the average reference
+        # directly – without going through a projector.
+        if 'eeg' in cfg.ch_types and cfg.eeg_reference == 'average':
+            epochs.set_eeg_reference('average')
+        else:
+            epochs.apply_proj()
+        epochs.pick(cfg.analyze_channels)
+
+    # We define the epochs and the labels
+    if isinstance(cfg.conditions, dict):
+        epochs_conds = [cfg.conditions[condition1],
+                        cfg.conditions[condition2]]
+        cond_names = [condition1, condition2]
+    else:
+        epochs_conds = cond_names = [condition1, condition2]
+        epochs_conds = [condition1, condition2]
+
+    # We have to use this approach because the conditions could be based on
+    # metadata selection, so simply using epochs[conds[0], conds[1]] would
+    # not work.
+    epochs = mne.concatenate_epochs([epochs[epochs_conds[0]],
+                                     epochs[epochs_conds[1]]])
+
+    n_cond1 = len(epochs[epochs_conds[0]])
+    n_cond2 = len(epochs[epochs_conds[1]])
+
+    X = epochs.get_data()
+    y = np.r_[np.ones(n_cond1), np.zeros(n_cond2)]
+
+    with config.get_parallel_backend():
+        classification_pipeline = make_pipeline(
+            Scaler(scalings='mean'),
+            Vectorizer(),  # So we can pass the data to scikit-learn
+            LogReg(
+                solver='liblinear',  # much faster than the default
+                random_state=cfg.random_state,
+                n_jobs=1,
+            )
+        )
+
+        # Now, actually run the classification, and evaluate it via a
+        # cross-validation procedure.
+        cv = StratifiedKFold(
+            shuffle=True,
+            random_state=cfg.random_state,
+            n_splits=cfg.decoding_n_splits,
+        )
+        scores = cross_val_score(
+            estimator=classification_pipeline,
+            X=X,
+            y=y,
+            cv=cv,
+            scoring='roc_auc',
+            n_jobs=1
+        )
+
+        # Save the scores
+        a_vs_b = f'{cond_names[0]}+{cond_names[1]}'.replace(op.sep, '')
+        processing = f'{a_vs_b}+FullEpochs+{cfg.decoding_metric}'
+        processing = processing.replace('_', '-').replace('-', '')
+
+        fname_mat = fname_epochs.copy().update(suffix='decoding',
+                                               processing=processing,
+                                               extension='.mat')
+        savemat(fname_mat, {'scores': scores})
+
+        fname_tsv = fname_mat.copy().update(extension='.tsv')
+        tabular_data = pd.Series(
+            {
+                'cond_1': cond_names[0],
+                'cond_2': cond_names[1],
+                'mean_crossval_score': scores.mean(axis=0),
+                'metric': cfg.decoding_metric
+            }
+        )
+        tabular_data = pd.DataFrame(tabular_data).T
+        tabular_data.to_csv(fname_tsv, sep='\t', index=False)
+
+
+def get_config(
+    subject: Optional[str] = None,
+    session: Optional[str] = None
+) -> SimpleNamespace:
+    cfg = SimpleNamespace(
+        task=config.get_task(),
+        datatype=config.get_datatype(),
+        acq=config.acq,
+        rec=config.rec,
+        space=config.space,
+        deriv_root=config.get_deriv_root(),
+        conditions=config.conditions,
+        contrasts=config.get_decoding_contrasts(),
+        decode=config.decode,
+        decoding_metric=config.decoding_metric,
+        decoding_n_splits=config.decoding_n_splits,
+        random_state=config.random_state,
+        analyze_channels=config.analyze_channels,
+        ch_types=config.ch_types,
+        eeg_reference=config.get_eeg_reference()
+    )
+    return cfg
+
+
+def main():
+    """Run time-by-time decoding."""
+    if not config.contrasts:
+        msg = 'No contrasts specified; not performing decoding.'
+        logger.info(**gen_log_kwargs(message=msg))
+        return
+
+    if not config.decode:
+        msg = 'No decoding requested by user.'
+        logger.info(**gen_log_kwargs(message=msg))
+        return
+
+    # Here we go parallel inside the :class:`mne.decoding.SlidingEstimator`
+    # so we don't dispatch manually to multiple jobs.
+    logs = []
+    for subject, session, (cond_1, cond_2) in itertools.product(
+        config.get_subjects(),
+        config.get_sessions(),
+        config.get_decoding_contrasts()
+    ):
+        log = run_epochs_decoding(
+            cfg=get_config(), subject=subject,
+            condition1=cond_1, condition2=cond_2,
+            session=session
+        )
+        logs.append(log)
+
+    config.save_logs(logs)
+
+
+if __name__ == '__main__':
+    main()