+++ scripts/source/__init__.py
@@ -2,12 +2,12 @@ from . import _01_make_bem_surfaces
 from . import _02_make_forward
 from . import _03_make_cov
 from . import _04_make_inverse
-from . import _05_group_average
+from . import _99_group_average
 
 SCRIPTS = (
     _01_make_bem_surfaces,
     _02_make_forward,
     _03_make_cov,
     _04_make_inverse,
-    _05_group_average,
+    _99_group_average,
 )